local_api/requirements.txt
flask
flask-restful
dataclasses_json
acryl-datahub